+++ README.md
@@ -36,4 +36,5 @@ Access database via console
 
 - [x] [Exercise 13.1-13.3](https://github.com/ruusukivi/fullstack-part13-postgresql/tree/13.1-13.3) -  Project setup and cli.js
 - [x] [Exercise 13.4](https://github.com/ruusukivi/fullstack-part13-postgresql/tree/13.4) -  index.js supporting GET, POST, DELETE from blogs
+- [x] [Exercise 13.5](https://github.com/ruusukivi/fullstack-part13-postgresql/tree/13.5) -  index.js supporting GET, POST, DELETE from blogs
