+++ util/config.js
@@ -1,0 +1,6 @@
+require('dotenv').config()
+
+module.exports = {
+  DATABASE_URL: process.env.DATABASE_URL,
+  PORT: process.env.PORT || 3001,
+}
