+++ util/db.js
@@ -1,0 +1,19 @@
+const Sequelize = require('sequelize')
+const { DATABASE_URL } = require('./config')
+
+const sequelize = new Sequelize(DATABASE_URL, {
+    dialect: 'postgres',
+});
+
+const connectToDatabase = async () => {
+    try {
+      await sequelize.authenticate()
+      console.log('database connected')
+    } catch (err) {
+      console.log('connecting database failed')
+      return process.exit(1)
+    }
+    return null
+}
+  
+module.exports = { connectToDatabase, sequelize }
